This commit seek to fix the following:
In addition, I am reorganizing the code and placing all relevent parts into their respective files and modules (this is should've been done at the beginning of the project).

1. The google books API keeps failing.
2. The four buttons in index.html are not redirecting properly, they are not showing the right records from the database. Fix this

Notes:
The restructure is done, however, there is an issue with the Blueprints
