+++ app/errors.py
@@ -1,0 +1,2 @@
+# insert error routes below
+
